Initial working pinentry config
The comments on how to set the environment variable for the shepherd gpg-agent
daemon is just for information. Will be deleted on the next commit since the
daemon does not need to be initialized to set that variable. Therfore
'SSH_AUTH_SOCK' can be set by the init script which calls shepherd.