Exposed INFOPATH to guix profiles by adding info-reader
In order to expose an enviroment variable to a guix profile there
needs to be a program that makes use of it (not only a package that leaves files
where those variables will search).

Eg.: 'stow' outputs a info file into the <profile>/share/info folder but
since the program itself won't make use of it, the 'INFOPATH' variable
exposed to the profile won't be updated. If the program 'info-reader'
is added, the 'INFOPATH' variable will be updated since the program
requires it to read info files. The same applies to other programs
such as 'man-db' with 'MANPATH' or 'python' with 'PYTHONPATH'.